requirements.txt
streamlit
onnxruntime
numpy
opencv-python
scikit-image
Pillow